pyproject.toml
[tool.poetry]
name = "pygeminador"
version = "0.1.0"
description = "Uma aplicação python com a IA Gemini do Google para ajudar no desenvolvimento de código"
authors = ["Lário Diniz <lario.developer@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
customtkinter = "^5.2.2"
google-generativeai = "^0.5.2"
python-dotenv = "^1.0.1"
pillow = "^10.3.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
blue = "^0.9.1"
pytest = "^8.2.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
